25.2. JOINing Across Multiple Relationships
  Joining Across Two Entities
Go open that up: src/Repository/QuestionRepository.php... here it is. To solve the N+1 problem, we joined directly across the previous q.tags relationship. Now we're going to need two joins to get to the tag table.

No problem: change q.tags to q.questionTags and alias that to question_tag. Then do an inner join from QuestionTag to Tag - ->innerJoin('question_tag.tag') - and alias that to tag.

  Cool! And we're still selecting the tag data... so that looks good to me.

  Refresh again and... another error! This one... is even more confusing.

    The parent object of entity result with alias tag was not found. The parent alias is question_tag.

  This is trying to say that it doesn't like that we're selecting the tag data... but we're not selecting the question_tag data that's between Question and Tag. Doing that is legal in SQL... but it messes up how Doctrine creates the objects, so it doesn't allow it.